[DataArray,Tag,MultiTag] Remove Mixins
