+++ nixio/pycore/multi_tag.py
@@ -7,7 +7,6 @@
 # LICENSE file in the root of the Project.
 from __future__ import (absolute_import, division, print_function)
 
-from ..multi_tag import MultiTagMixin
 from .tag import BaseTag
 from .data_array import DataArray
 from .data_view import DataView
@@ -15,8 +14,10 @@ from ..link_type import LinkType
 from .exceptions import (OutOfBounds, IncompatibleDimensions,
                          UninitializedEntity)
 
+from .tag import ReferenceProxyList, FeatureProxyList
 
-class MultiTag(BaseTag, MultiTagMixin):
+
+class MultiTag(BaseTag):
 
     def __init__(self, nixparent, h5group):
         super(MultiTag, self).__init__(nixparent, h5group)
@@ -201,3 +202,35 @@ class MultiTag(BaseTag, MultiTagMixin):
         count = da.data_extent
         offset = (0,) * len(count)
         return DataView(da, count, offset)
+
+    @property
+    def references(self):
+        """
+        A property containing all data arrays referenced by the tag. Referenced
+        data arrays can be obtained by index or their id. References can be
+        removed from the list, removing a referenced DataArray will not remove
+        it from the file. New references can be added using the append method
+        of the list.
+        This is a read only attribute.
+
+        :type: RefProxyList of DataArray
+        """
+        if not hasattr(self, "_references"):
+            setattr(self, "_references", ReferenceProxyList(self))
+        return self._references
+
+    @property
+    def features(self):
+        """
+        A property containing all features of the tag. Features can be obtained
+        via their index or their id. Features can be deleted from the list.
+        Adding new features to the tag is done using the create_feature method.
+        This is a read only attribute.
+
+        :type: ProxyList of Feature.
+        """
+        if not hasattr(self, "_features"):
+            setattr(self, "_features", FeatureProxyList(self))
+        return self._features
+
+
